feat: move map-to-g1 to swbls12381 package
